+++ lib/presentation/utils/ui_helpers.dart
@@ -22,12 +22,23 @@ Color darkenColor(Color color, {double factor = 0.2}) {
   return Color.fromRGBO(red, green, blue, 1);
 }
 
+String getEmojiForRefreshRate(double refreshRate) {
+  if (refreshRate >= 120) {
+    return '🚀'; // High refresh rate, indicating very smooth
+  } else if (refreshRate >= 90) {
+    return '✨'; // Moderately high, indicating good smoothness
+  } else {
+    return '🔋'; // Standard refresh rate
+  }
+}
+
 class AppColors {
   static const Color primaryColor = Color(0xFF0D47A1);
   static const Color secondaryColor = Color(0xFF2196F3);
   static const Color tertiaryColor = Color(0xFF00BFA5);
   static const Color quaternaryColor = Color(0xFF7f4bd7);
   static const Color quinaryColor = Color(0xFF3F51B5);
+  static const Color senaryColor = Color(0xFFD49114);
   //indigo
   static const Color eventCardColor = Color(0xFF3F51B5);
   //deep blue
