+++ lib/presentation/widgets/event_card.dart
