+++ lib/presentation/widgets/login_button.dart
@@ -1,0 +1,27 @@
+import 'package:auto_route/auto_route.dart';
+import 'package:flutter/material.dart';
+import 'package:project_kepler/core/extensions/build_context_ext.dart';
+
+class LoginButton extends StatelessWidget {
+  const LoginButton({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return ElevatedButton(
+      style: ElevatedButton.styleFrom(
+        backgroundColor: context.theme.colorScheme.primary,
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(10),
+        ),
+      ),
+      onPressed: () {
+        context.router.pushNamed('/login');
+      },
+      child: Text(context.l10n.logIn,
+          style: context.theme.textTheme.labelLarge
+              ?.copyWith(color: Colors.white)),
+    );
+  }
+}
